1 引入模块化的思路，对各个功能模块进行解耦 2 引入connd进程的支持，从而可以灵活地处理跳服的场景 3 机器人进行修改，支持通过send发送消息，加入msgid枚举进lua环境

+++ inc/common/shm/shm_pool_inl.h
@@ -293,7 +293,7 @@ template<class T>
 BOOL CShmObject<T>::init(int32_t shm_type, BOOL is_resume)
 {
 	SHM_POOL* pool = NULL;
-	uint64_t pool_size = sizeof(SHM_POOL) + sizeof(SHM_UNIT_DATA<T>);
+	int32_t pool_size = sizeof(SHM_POOL) + sizeof(SHM_UNIT_DATA<T>);
 
 	LOG_PROCESS_ERROR_DETAIL(shm_type > stdInvalid, "shm_type is %d", shm_type);
 	LOG_PROCESS_ERROR_DETAIL(shm_type < stdTotal, "shm type is %d", shm_type);
